update styles
Add styles for the title bar to conform with the rest of the design language including margin, border-radius, and opacity.  Modify style rules for bookmark, adress, and tab bars to match.

NOTE: Only opacity style rules take effect on a seperated settings window.

Tested layouts with bookmarks, address, and tabs on top and bottom and hidden..